YAN-1823: The first release of the directory_watcher code and the start of related tests.

See merge request ska-telescope/ska-dlm-client!8

+++ src/ska_dlm_client/__init__.py
@@ -17,14 +17,19 @@ DLM_CLIENT_HOME = Path.home() / ".dlm_client/"
 """The configuration path of dlm-client."""
 
 
-def read_config(user_config_file: Path = DLM_CLIENT_HOME / "config.yaml") -> benedict:
+def read_config(user_config_file: Path = DLM_CLIENT_HOME / "config.yaml") -> benedict | None:
     """Read the config file and return the config dictionary."""
     if not user_config_file.exists():
         # create the default user config in DLM_CLIENT_HOME if it does not already exist
         print(f"DLM client config file {user_config_file} not found - creating and using default")
         user_config_file.parent.mkdir(exist_ok=True)
         default_user_config_file = DLM_CLIENT_LIB_DIR / "config.yaml"
-        shutil.copy(default_user_config_file, user_config_file)
+        # TODO: Fix for docker compose build to correctly copy default file to package!
+        if Path(default_user_config_file).exists():
+            shutil.copy(default_user_config_file, user_config_file)
+        else:
+            print(f"Default DLM client config {default_user_config_file} does not exit, ignoring.")
+            return None
 
     with open(user_config_file, "r", encoding="utf-8") as file:
         return benedict(yaml.safe_load(file))
